requirements.txt
anthropic>=0.40.0
flask>=3.0.0
flask-socketio>=5.3.6
aiohttp>=3.9.0
pyyaml>=6.0
numpy>=1.26.0
opencv-python-headless>=4.9.0
dlib>=19.24.0
face-recognition>=1.3.0

# Ecobee integration
aiohttp>=3.9.0  # already included above
